Moved maximumFileSize from common configuration parameters to hdfs configuration, reducing encapsulation in PartitionFileImpl. Refactored all tests according to the change.
