+++ src/main/java/com/teragrep/cfe_39/configuration/CommonConfiguration.java
@@ -111,31 +111,6 @@ public final class CommonConfiguration {
         return config.getOrDefault("queueDirectory", System.getProperty("user.dir") + "/rpm/resources/queue");
     }
 
-    public long maximumFileSize() {
-        final String numString = config.get("maximumFileSize");
-        if (numString == null) {
-            throw new ConfigurationException("Configuration error. <maximumFileSize> must be set.");
-        }
-        else {
-            final long maximumFileSize;
-            try {
-                maximumFileSize = Long.parseLong(numString);
-            }
-            catch (NumberFormatException e) {
-                LOGGER.error("Configuration error. Invalid value for <maximumFileSize>: <{}>", e.getMessage());
-                throw new RuntimeException(e);
-            }
-            if (maximumFileSize <= 0) {
-                throw new ConfigurationException(
-                        "Configuration error. <maximumFileSize> must be a positive long value."
-                );
-            }
-            else {
-                return maximumFileSize;
-            }
-        }
-    }
-
     public boolean skipNonRFC5424Records() {
         final String skipString = config.get("skipNonRFC5424Records");
         if (skipString == null) {
