+++ src/test/java/com/teragrep/cfe_39/CommonConfigurationTest.java
@@ -70,7 +70,7 @@ public class CommonConfigurationTest {
             map = pathConfiguration.asMap();
             Assertions
                     .assertEquals(
-                            "{numOfConsumers=2, queueTopicPattern=^testConsumerTopic-*$, queueDirectory=/opt/teragrep/cfe_39/etc/AVRO/, skipNonRFC5424Records=true, maximumFileSize=3000, skipEmptyRFC5424Records=true, consumerTimeout=600000}",
+                            "{numOfConsumers=2, queueTopicPattern=^testConsumerTopic-*$, queueDirectory=/opt/teragrep/cfe_39/etc/AVRO/, skipNonRFC5424Records=true, skipEmptyRFC5424Records=true, consumerTimeout=600000}",
                             map.toString()
                     );
             CommonConfiguration commonConfig = new CommonConfiguration(map);
@@ -83,7 +83,6 @@ public class CommonConfigurationTest {
             Assertions
                     .assertEquals(System.getProperty("user.dir") + "/rpm/resources/log4j2.properties", commonConfig.log4j2ConfigurationFile());
             Assertions.assertEquals(2, commonConfig.numOfConsumers());
-            Assertions.assertEquals(3000, commonConfig.maximumFileSize());
             Assertions.assertEquals(600000, commonConfig.consumerTimeout());
             Assertions.assertTrue(commonConfig.skipNonRFC5424Records());
             Assertions.assertTrue(commonConfig.skipEmptyRFC5424Records());
