+++ src/test/resources/failProcessing.application.properties
@@ -10,8 +10,6 @@ queueTopicPattern=^testConsumerTopic-*$
 numOfConsumers=2
 # Directory where AVRO files are constructed for HDFS
 queueDirectory=/opt/teragrep/cfe_39/etc/AVRO/
-# The maximum file size for AVRO-files that are to be stored in HDFS database.
-maximumFileSize=3000
 # Boolean for deciding if records not in RFC5424 should be skipped or not.
 skipNonRFC5424Records=false
 # Boolean for deciding if empty RFC5424 records should be skipped or not.
