+++ src/test/resources/valid.hdfs.properties
@@ -16,3 +16,5 @@ dfs.client.use.datanode.hostname=false
 hadoop.kerberos.keytab.login.autorenewal.enabled=true
 dfs.data.transfer.protection=test
 dfs.encrypt.data.transfer.cipher.suites=test
+# The maximum file size for AVRO-files that are to be stored in HDFS database.
+maximumFileSize=3000
